+++ lib/compressor.dart
@@ -1,0 +1,2 @@
+export 'compression_result.dart';
+export 'light_compressor.dart';